cron_requirement.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pytz>=2023.3